refactor structure, add workspaceview component
