+++ src/Webapp/src/components/admin/index.ts
@@ -1,4 +1,4 @@
-import WorkspaceTable from './workspace/WorkspaceTable'
-import UserSettings from './UserSettings'
+import UserSettings from './settings/UserSettings'
+import WorkspaceView from './workspace/WorkspaceView'
 
-export { UserSettings, WorkspaceTable }
+export { UserSettings, WorkspaceView }
