+++ src/Webapp/src/components/admin/workspace/WorkspaceView.tsx
@@ -1,0 +1,14 @@
+import { FC } from 'react'
+
+import WorkspaceTable from './WorkspaceTable'
+
+const WorkspaceView: FC = () => {
+  return (
+    <>
+      <h1>Рабочие пространства</h1>
+      <WorkspaceTable />
+    </>
+  )
+}
+
+export default WorkspaceView
